pyproject.toml
[tool.poetry]
name = "sell-project"
version = "0.1.0"
description = ""
authors = ["Bogdan <bogdan.vp2015@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
setuptools = "^69.1.1"
django = "^5.0.3"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
elasticsearch = "^7.8.1"
elasticsearch-dsl = "^7.3.0"
django-elasticsearch-dsl = "^7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
